requirements.txt
# Core LLM & API
openai>=1.30.0

# Retrieval
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
sentence-transformers>=2.7.0

# Evaluation
deepeval>=0.21.0
ragchecker>=0.1.0

# Data & utilities
numpy>=1.26.0
tqdm>=4.66.0
python-dotenv>=1.0.0
requests>=2.31.0
